pyproject.toml
[project]
name = "agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "google-auth>=2.41.1",
    "google-cloud-aiplatform>=1.122.0",
    "google-genai>=1.46.0",
    "google-generativeai>=0.8.5",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "playwright>=1.55.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.4.4",
    "pypdf>=5.1.0",
    "pykakasi>=2.3.0",
    "jaconv>=0.4.0",
]

